awk: bugfix: autoload on sys:awk-state.
The problem is that compiled code which uses the awk macro
won't load because it invokes make-struct for the unknown
sys:awk-state type. Interpreted code doesn't have the problem
because it has to expand the awk macro from scratch to
generate the make-struct call, and the awk macro triggers
the autoload.

* lisplib.c (awk_set_entries): Add sys:awk-state to autoload
list for awk.tl.