compiler: fix broken block*.
Compiling a block* fails with an exception: nil is accessed
as a structure.

* share/txr/stdlib/compiler.tl (comp-block): when compiling
block*, nenv is nil; we must use env when compiling
the name subexpression. Also, we can't use nenv when compiling
the body. We must use nenv when it is available (block case)
or else fall back on env (block* case).