awk: bugfix: block optimized away by compiler.
This breaks the (next) awk macro, breaking awk expressions
which want to short-circuit to the next record.

* share/txr/stdlib/awk.tl (sys:awk-state loop): The :awk-rec
block is being optimized away by the compiler because it
doesn't contain any direct function calls to functions that
are not in the standard library. We must use block*, which
isn't subject to this optimization.