defvar: warn about prior lexical uses.
Now that the compiler has a more efficient treatment of global
lexical variables, code which accesses global variables that
have not yet been defined will misbehave if the intent is to
for those variables to be dynamically scoped.

There is such a bug in the op expander, in fact.

* eval.c (me_def_variable): When defvar/defparm are expanding,
they now check whether there is an outstanding unbound warning
against the variable. If so, then a warning is issued that the
variable was previously used lexically and is now being marked
special.

* unwind.c (uw_warning_exists): New function.

* unwind.h (uw_warning_exists): Declared.