listener: avoid unnecessary string duplication.
* parser.c (repl): There is no need to use chk_strdup on the
string inside histfile. We can just use the original string,
since it won't be garbage collected. The existing
gc_hint(histfile) at the end of the function ensures this.
The reason the chk_strdup was done is that originally this
was a utf8_dup_to that I just blindly replaced when the
listener Unicode conversion took place.