compile: handle functions that have environments.
With this patch, the compile function can handle interpreted
function objects that have captured environments.

For instance, if the following expression is evaluated

  (let ((counter 0))
    (labels ((bm () (bump))
             (bump () (inc counter)))
      (lambda () (bm))))

then a function object emerges. We can now feed this
function object to the compile function; the environment
will now be handled.

Of course, the above expression is already compileable;
compile-toplevel handles it and so does the file compiler.

This patch allows the expression to be interpreted and then
the function object to be compiled, without access to the
surrounding expression. The compiled function will contain a
compiled version of the environment, carrying compiled
versions of the captured variables and their contents.

* eval.c (env_vbindings, env_fbindings, env_next): New static
functions.
(eval_init): Register env-vbinding, env-fbindings and env-next
intrinsics.

* share/txr/stdlib/compiler.tl (sys:env-to-let): New function.
(usr:compile): Wrap the interpreted lambda terms with let
bindings carefully reconstructed from their captured
environments.

* txr.1: Documented new intrinsic functions.