case macros: bugfix empty case.
* eval.c (me_case): When there are no keys, then it is
logically true that all keys are integer, and the hash table
logic kicks in. The minkey and maxkey variables are supposed
to be calculated as zero in that case, but the empty test is
bungled since nkeys doesn't test false when it is zero.  We
end up with minkey and maxkey containing nil which get passed
to the minus function. And so, here we fix the empty test.