defvar: bugfix: bad state caused by exception.
The following situation can occur:

  (defvar v expr) ;; expr throws!

What happens is that the internal hash table of global
variables ends up with an entry that has a nil value,
instead of the expected (v . <value>) cons cell.
The hash table entry is created when the table is probed
for the existence of v, but then is never populated
with a binding because expr throws.

The consequence then is that (boundp v) returns nil, yet
subsequent (defvar v ...) expressions think that v exists, and
refuse to define it. (defparm v ...) also fails; it relies on
(defvar ...) which fails to define the variable, and then
tries to assign to it, which throws.

* eval.c (rt_defvarl): If the hash table exists, but has a nil
value, treat that as an undefined variable.  Thus, define the
variable not only if the hash cell is newly made, but also if
it already exists, with a null cdr.