Get alloca from stdlib.h, if possible.
* configure: try <stdlib.h> first for alloca. This should fix
a build issue which happens on the Musl library and perhaps
elsewhere. The problem on Musl is that #include <stdlib.h>
already includes <alloca.h>. That header contains an alloca
macro which interfers with our subsequent #include ALLOCA_H
directive: ALLOCA_H expands to <alloca.h> and the alloca token
gets further expanded.