asm/compiler: rename small level/index constants.
* share/txr/stdlib/vm-param.tl (%max-sm-lev-idx%): This
constant is named inconsistently relative to %max-lev-idx%.
It is providing the maximum level (encodable in a small
operand), whereas %max-lev-idx% provides the maximum index
within a level. It is hereby renamed to %max-sm-lev%.
The %max-sm-lev-idx% name is re-used to denote the quantity
which it suggests: the maximum index within a level
(encodable in a small operand), which is 63.

* share/txr/stdlib/asm.tl (small-op-p): Use the new
%max-sm-lev-idx% in place of %sm-lev-size%, getting rid of the
funny range starting with -1. Replace the original
%max-sm-lev-idx% with its new name, %max-sm-lev%.