carray: fix vec/list conversion bug.
If a zero-length carray is converted with vec-carray or
list-carray and the null-term-p argument is t, there is an
exception about a negative index. An empty vector or list
should be returned in this case, and the documentation says
exactly that.

Also, if a carray of unknown length is converted, there is
an exception from vec-carray, as documented, but it's
an uninformative one that is incidentally produced when -1
is passed to the vec function. The list-carray just returns
nil, contravening the documentation.

* ffi.c (vec_carray, list_carray): Fix the problems described
above.

* txr.1: Reviewing the documentation for these functions, an
improperly terminated sentence was found.