configure: improvements related to int_ptr_t.
* configure: We remove the uintptr variable and do not
generate the HAVE_UINTPTR_T constant in config.h. This is
always 1, and not actually tested anywhere. In situations when
uintptr is not established, intptr is also not established and
the configure script fails. We simply assume that for whatever
type we detect as inptr_t, we can derive the unsigned type.
(INT_PTR_MAX): We define this a bit differently; instead of
interpolating into the expression the underlying C type, we
use the int_ptr_t typedef that the previous lines of config.h
establish.
(UINT_PTR_MAX): New constant introduced in config.h.