itypes: take advantage of double_intptr_t.
If double_intptr_t is at least 64 bits wide (which is always
going to be true) we can take advantage of it.

* itypes.c (c_i64, c_u64): Provide implementations that use
the new dbl_intptr_t and dbl_uintptr_t infrastructure.
The fallback implementation that uses ash and logior is
basically not used now.
(num_64, unum_64): New functions for obtaining a Lisp value
from a 64 bit number, abstracting whether this is done via the
cnum or dbl_cnum route.

* itypes.h (u64_t, i64_t): Define in terms of double_uintptr_t
and double_intptr_t, if no other 64 bit type is available.
(num_64, unum_64): Declared.