streams: use Boolean return value for stdio_fseek.
* stream.c (stdio_seek): Handle new-style Boolean return from
stdio_fseek.

* sysif.c (stdio_fseek): Return an int indication that is 1
for success, 0 for failure. There was a mistaken assumption
here that fseeko returns the file offset, and the return value
in the fseek case was mistakenly harmonized, using a wasteful
call to stdio_ftell. The only call to this function relies
only on a Boolean success/fail indication.

* sysif.h (stdio_fseek): Declaration updated.