command line: support -f in Hash Bang Null Hack.
* txr.c (txr_main): In processing the -f option, if there is
already an input source, but its name exactly matches the
argument of the -f option, then ignore the -f and its option
intead of terminating with a diagnostic. This allows the user
to use -f in the Null Hack to achieve the usual trick of
allowing the script to take options passed through to the
interpreter.

* txr.1: Documented.