@(next): bugfix: nothrow not causing failed match.
* match.c (v_next_impl): Contrary to the documentation and to
"classic" TXR behavior exhibited in compatibility mode,
when @(next "file" :nothrow) fails to open "file", it treats
the situation as a cue to read from standard input, rather
than to treat the situation as a failed match. This is because
complex_open returns nil, and when that is passed to
lazy_stream_cons, that function defaults to standard input.