linenoise: bugfix: regression in mlmode line wrap.
The following problem happens: when charaters are inserted
past the end of the line such that it wraps, hitting backspace
or any cursor movement causes a spurious scroll.

This was caused on Nov 1 2018 by bf85503b (linenoise: avoid
refresh for new text in multi-line mode).

The reason is that the maxrows variable isn't updated when
we trivially add a character without repainting.

* linenoise/linenoise.c (lino state): Document special value
for need_refresh: when it is 2, the refresh doesn't perform a
any output, but recalculates maxrows.
(refreh_multiline): If need_refresh is 2, bail after
updating maxrows.
(edit_insert): When trivially adding a character at the
end and just outputting it, if in multi-line mode, set
need_refresh to 2.