ffi: closure: clear stale saved exit point.
What happens if a FFI closure intercepts an exception, returns
to the foreign code, but that caller does not return to Lisp
where the exception will continue? Suppose the caller invokes
another closure.  In that situation, the lingering value of
the s_exit_point variable causes problems for FFI calls.

An example of this is callbacks from an event dispatching
framework such as a GUI. When these callback closures return,
control returns to the event loop, which itself doesn't
immediately return to Lisp code. Rather, it processes more
events and invokes more callbacks.

* ffi.c (ffi_closure_dispatch_safe):  One thing we can do
is clear s_exit_point whenever we dispatch a closure.