structs: fix missing lnew operator.
* lisplib.c (struct_set_entries): Add missing entry for lnew
in the name array. This is not just a missing auto-load issue.
Because we don't intern the symbol, struct.tl ends up defining
a sys:lnew macro, and not usr:lnew. The symbol usr:lnew
doesn't exist and so when application code tries to use lnew,
it's trying to invoke pub:lnew.