float: turn out-of-range calculations into exceptions.
On platforms that have the C99 fpclassify, we can use it to
banish infinity and NaN representations. If such a thing
arises, we throw an exception rather than producing an object
that prints as #<bad-float>.

* configure: add detection for fpclassify.

* lib.c (bad_float): New inline function and macro.
(flo): If the argument is other than zero, a normal
value or a subnormal, then throw an exception.  We thereby
refuse to admit such objects into our numeric object
system.