listener: ensure history and temp files are rw-------.
For security, the temporary files used by the "edit in
external editor" feature of the listener, as well as the
listener history file, should be readable and writable only to
the owner.

This relates to Debian bug 832460 against the Linenoise
library:

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=832460

In the TXR fork of the linenoise library, since we have an OS
abstraction invoked by callback functions, we fix this
entirely outside of linenoise.

I don't agree with the upstream approach of fiddling with the
umask and doing a chmod on the path.

Since we are truncating and overwriting the file, all we
have to do is, before writing any data, fchmod it to the
required permissions.

* parser.c (lino_open): If the file is being open for
overwriting, then let's set its permissions so that it's
readable and writable for the user only.