tagbody: code replication bug.
tagbody translates to a loop around a sys:switch special form.
The compiler recognizes substructure sharing among the
branches of the the sys:switch code and de-duplicates the
generated assembly code accordingly. Unfortunately, the
expander for sys:switch has a bug which spoils the
substructure sharing.

* eval.c (expand_forms_ss): We must process the incoming form
list itself into the hash, not just that list's cdr.
Onthe other hand, there is no need to hash the cars of the
forms list (the forms themselves), so the code simplifies
quite a bit.